bug: fix leading/trailing punctuation highlights.
Prior to this commit leading/trailing punctuation was removed prior to
calculating the highlight mask. Instead, we should only remove
punctuation before consulting the spellcheck API for correctness or
suggestions. This lets the masking code mask the entire section of text,
including punctuation, but doesn't include leading/trailing punctuation
when consulting the dictionary.