project: prefix Blightspell src files
In the context of a Blightmud plugin it appears as though `require`'ing
a path like "src/settings" can result in loading a _different_
plugin's `src/settings.lua` file.

To avoid this clash, this commit prefixes all of the Blightspell Lua
files with `blightspell_` and updates the `main.lua` `require`
statements accordingly.

This allows Blightspell to co-exist with a plugin like `blightmud_mcp`
that already has a `src/settings.lua` file. Without this change we see
mysterious settting init failure because not all of the Blightspell code
is loaded as expected!